feat(GenericContainer): introduce LabelSetting trait for managing container labels
